chore: Addressed PR feedback.
- Moved dependency resolution functions to dependency_resolver.py
- Moved find_valid_url to repo_finder to avoid circular dependencies
- Added URL for deps.dev API documentation as comment
